# libct/system.Stat: fix panic, speed up, add benchmarks