#### 1.    libct/system/proc_test: fix, improve, add benchmarks
    
1. Add a test case that tests parentheses in command.
    
2. Replace individual comparisons with `reflect.DeepEqual`.
   This also fixes wrong `%`-style types used in `Fatalf` statements.
    
3. Replace `Fatalf` with `Errorf` so we don't bail out on the first
   failure, and do not check result on error.
    
4. Add two benchmarks.

#### 2. libct/system.Stat: fix/improve/speedup

1. Remove PID field. It is totally useless -- we already know the PID.
    
2. Rewrite `parseStat()` to make it faster and more correct:
   - do not use `fmt.Scanf` as it is very slow;
   - avoid splitting data into 20+ fields, of which we only need 2 (should also help garbage collector);
   - make sure to not panic on short lines and other bad input;
   - add some bad input tests (some fail with old code);
   - use `LastIndexByte` instead of `LastIndex` where appropriate.
 
Benchmarks:

before (from the previous commit message):

```    
BenchmarkParseStat-4              116415             10804 ns/op
BenchmarkParseRealStat-4             240           4781769 ns/op
```
    
after:
```    
BenchmarkParseStat-4           1164948              1068 ns/op
BenchmarkParseRealStat-4           331           3458315 ns/op
```    

We are seeing 10x speedup in a synthetic benchmark, and about 1.4x
speedup in a real world benchmark. All this while being more scrupulous
about any possible errors. I also suspect much less allocations and thus
garbage to collect.
     
While at it, do not ignore any possible errors, and properly wrap those.

3. Add `I` and `P` process states (available since Linux 4.14).

## History
    
* v2: use pkg/errors more, remove t.Logf from test
* v3: rebased; drop pkg/errors; gofumpt'ed
* v4: rebased; improved description
* v5: rebased; mention bad input tests, added second benchmark results
* v6: remove PID field, do not use strings.Split, further speedup.
* v7: readability improvements, added new process states.
* v8: improved parser to bail earier on incorrect data, more doc, more tests.