@kolyshkin 2020-12-01
Rebased to benefit from #2690 (while Travis was not able to test this PR for 6+ hours :cry:)