@kolyshkin 2021-01-06
rebased (for new CI)