@cyphar 2021-02-04
Is this a bugfix or an improvement? If it's an improvement it belongs in post-1.0 as a milestone.