@cyphar 2021-02-05
Yeah agreed, this change looks a little scary tbh.