@thaJeztah 2021-06-11
linter is failing;

```
 Error: File is not `gofumpt`-ed (gofumpt)
```