@kolyshkin 2021-06-24
v3: rebased; drop using pkg/errors; gofumpt'ed