@kolyshkin 2021-08-04
Some quite interesting observations:
* removing `strings.Split` (which result in ~20 fields slice) and a couple of `fmt.Scanf`s results in 7x improvement!
* removing `strings.Split` (for much shorter data, only the first two fields) results in further 2x improvement!